+++ src/pages/ProfileEditPage/hooks/useFileUpload.ts
@@ -1,4 +1,14 @@
+import { useState } from "react"
+
+import { useMutation } from "@tanstack/react-query"
+
+import { postFile } from "@apis/file/postFile"
+
 const useFileUpload = () => {
+  const [isDragging, setIsDragging] = useState(false)
+  const [fileBase64, setFileBase64] = useState<string | undefined>()
+  const [responsedFileUrl, setResponsedFileUrl] = useState("")
+
   const convertFileToBase64 = (file: File) => {
     return new Promise<string>((resolve) => {
       const reader = new FileReader()
@@ -12,7 +22,76 @@ const useFileUpload = () => {
     })
   }
 
-  return { convertFileToBase64 }
+  const uploadFileMutation = useMutation({
+    mutationFn: async (file: File) => await postFile(file),
+    onSuccess(data) {
+      setResponsedFileUrl(data?.fileUrl)
+      console.log(data)
+    },
+  })
+
+  const handleFileChange = async (
+    event: React.ChangeEvent<HTMLInputElement>,
+  ) => {
+    if (event.target.files && event.target.files.length > 0) {
+      const selectedFile = event.target.files[0]
+      const base64 = await convertFileToBase64(selectedFile)
+      setFileBase64(base64)
+      console.log(selectedFile)
+
+      uploadFileMutation.mutate(selectedFile)
+    }
+  }
+
+  const handleDragEnter = (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault()
+    event.stopPropagation()
+
+    if (event?.dataTransfer?.items && event.dataTransfer.items.length > 0) {
+      setIsDragging(true)
+    }
+  }
+
+  const handleDragLeave = (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault()
+    event.stopPropagation()
+    setIsDragging(false)
+  }
+
+  const handleDragOver = (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault()
+    event.stopPropagation()
+  }
+
+  const handleFileDrop = async (event: React.DragEvent<HTMLDivElement>) => {
+    event.preventDefault()
+    event.stopPropagation()
+
+    const files = event?.dataTransfer?.files
+
+    if (files && files?.length > 0) {
+      const selectedFile = files[0]
+
+      const base64 = await convertFileToBase64(selectedFile)
+      setFileBase64(base64)
+
+      uploadFileMutation.mutate(selectedFile)
+    }
+    setIsDragging(false)
+  }
+
+  return {
+    uploadFileMutation,
+    convertFileToBase64,
+    handleFileChange,
+    handleDragEnter,
+    handleDragLeave,
+    handleDragOver,
+    handleFileDrop,
+    isDragging,
+    fileBase64,
+    responsedFileUrl,
+  }
 }
 
 export default useFileUpload
